Spam protection + update to newer django specs.
in `punkweb/rest/views.py`, `UserCreateView`'s `post` function now adds a "fingerprint" to a list. When another call comes through, it cross-references that list with the new requests' ip. If it is in the list, it checks if it has been > two minutes. If it has, it removes it from the list and program flow continues as usual. If it hasn't, it returns `401 unauthorized`.

Added  `409` status code if `UserCreateSerializer` does not return a valid `serializer`.